Allow forge name & password of initial account for RPM (like it was done for the source install).

+++ src/fusionforge.spec
@@ -561,8 +561,14 @@ if [ "$1" -eq "1" ]; then
 	if [ $ret -ne 0 ] ; then
 	    FFORGE_DB=%{dbname}
 	    FFORGE_USER=%{dbuser}
-	    FFORGE_ADMIN_USER=%{fforge_admin}
-	    FFORGE_ADMIN_PASSWORD=$(/bin/dd if=/dev/urandom bs=32 count=1 2>/dev/null | /usr/bin/sha1sum | cut -c1-8)
+	    if [ "x${FFORGE_ADMIN_USER}" = "x" ]
+	    then
+	        FFORGE_ADMIN_USER=%{fforge_admin}
+	    fi
+	    if [ "x${FFORGE_ADMIN_PASSWORD}" = "x" ]
+	    then
+	        FFORGE_ADMIN_PASSWORD=$(/bin/dd if=/dev/urandom bs=32 count=1 2>/dev/null | /usr/bin/sha1sum | cut -c1-8)
+	    fi
 	    export FFORGE_DB FFORGE_USER FFORGE_ADMIN_USER FFORGE_ADMIN_PASSWORD
 	    /usr/bin/php %{FORGE_DIR}/fusionforge-install-3-db.php >>/var/log/%{name}-install.log 2>&1
 	else
